Update cran-comments.md
Added summary:

## Summary of checks in all environments
Apart from 6 file size NOTES (package size is 5.1 to 5.8 Mb depending on environment), and failures on 2 environments which do not appear to be among the flavors tested on CRAN, and 1 on which required packages were not available, the package passed all checks on the remaining 12 environments. The file size has been considerably reduced from about 16.7 Mb by moving vignettes to a separate online book, referenced in this package, so I request an exception be made to the 5 Mb rule, as further reduction would damage the integrity of the package. The check details on all 21 environments follows.